refactor: extract Synchronization to cryptarchia-sync crate
