+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
   "clients/common-http-client",
   "clients/executor-http-client",
   "consensus/cryptarchia-engine",
+  "consensus/cryptarchia-sync/core",
   "consensus/cryptarchia-sync/network",
   "ledger/nomos-ledger",
   "nodes/nomos-executor",
@@ -55,6 +56,7 @@ cl                       = { path = "./nomos-core/cl" }
 common-http-client       = { path = "./clients/common-http-client" }
 cryptarchia-consensus    = { path = "./nomos-services/cryptarchia-consensus" }
 cryptarchia-engine       = { path = "./consensus/cryptarchia-engine" }
+cryptarchia-sync         = { path = "./consensus/cryptarchia-sync/core" }
 cryptarchia-sync-network = { path = "./consensus/cryptarchia-sync/network" }
 executor-http-client     = { path = "./clients/executor-http-client" }
 full-replication         = { path = "./nomos-da/full-replication" }
