+++ dependencies_graph.dot
@@ -13,35 +13,36 @@ digraph {
     11 [ label = "nomos-executor" shape = box]
     12 [ label = "nomos-api" shape = box]
     13 [ label = "cryptarchia-consensus" shape = box]
-    14 [ label = "cryptarchia-sync-network" shape = box]
-    15 [ label = "subnetworks-assignations" shape = box]
-    16 [ label = "nomos-blend-service" shape = box]
-    17 [ label = "nomos-blend" shape = box]
-    18 [ label = "nomos-blend-message" shape = box]
-    19 [ label = "nomos-blend-network" shape = box]
-    20 [ label = "nomos-libp2p" shape = box]
-    21 [ label = "nomos-network" shape = box]
-    22 [ label = "services-utils" shape = box]
-    23 [ label = "nomos-da-sampling" shape = box]
-    24 [ label = "nomos-da-network-core" shape = box]
-    25 [ label = "nomos-da-network-service" shape = box]
-    26 [ label = "nomos-tracing" shape = box]
-    27 [ label = "nomos-da-storage" shape = box]
-    28 [ label = "nomos-da-verifier" shape = box]
-    29 [ label = "nomos-storage" shape = box]
-    30 [ label = "nomos-mempool" shape = box]
-    31 [ label = "nomos-tracing-service" shape = box]
-    32 [ label = "nomos-time" shape = box]
-    33 [ label = "nomos-da-dispersal" shape = box]
-    34 [ label = "nomos-da-indexer" shape = box]
-    35 [ label = "nomos-node" shape = box]
-    36 [ label = "nomos-system-sig" shape = box]
-    37 [ label = "bundler" shape = box]
-    38 [ label = "nomos-cli" shape = box]
-    39 [ label = "full-replication" shape = box]
-    40 [ label = "key-management-system" shape = box]
-    41 [ label = "cfgsync" shape = box]
-    42 [ label = "tests" shape = box]
+    14 [ label = "cryptarchia-sync" shape = box]
+    15 [ label = "cryptarchia-sync-network" shape = box]
+    16 [ label = "subnetworks-assignations" shape = box]
+    17 [ label = "nomos-blend-service" shape = box]
+    18 [ label = "nomos-blend" shape = box]
+    19 [ label = "nomos-blend-message" shape = box]
+    20 [ label = "nomos-blend-network" shape = box]
+    21 [ label = "nomos-libp2p" shape = box]
+    22 [ label = "nomos-network" shape = box]
+    23 [ label = "services-utils" shape = box]
+    24 [ label = "nomos-da-sampling" shape = box]
+    25 [ label = "nomos-da-network-core" shape = box]
+    26 [ label = "nomos-da-network-service" shape = box]
+    27 [ label = "nomos-tracing" shape = box]
+    28 [ label = "nomos-da-storage" shape = box]
+    29 [ label = "nomos-da-verifier" shape = box]
+    30 [ label = "nomos-storage" shape = box]
+    31 [ label = "nomos-mempool" shape = box]
+    32 [ label = "nomos-tracing-service" shape = box]
+    33 [ label = "nomos-time" shape = box]
+    34 [ label = "nomos-da-dispersal" shape = box]
+    35 [ label = "nomos-da-indexer" shape = box]
+    36 [ label = "nomos-node" shape = box]
+    37 [ label = "nomos-system-sig" shape = box]
+    38 [ label = "bundler" shape = box]
+    39 [ label = "nomos-cli" shape = box]
+    40 [ label = "full-replication" shape = box]
+    41 [ label = "key-management-system" shape = box]
+    42 [ label = "cfgsync" shape = box]
+    43 [ label = "tests" shape = box]
     0 -> 7 [ ]
     1 -> 5 [ ]
     3 -> 4 [ style = dotted]
@@ -53,50 +54,52 @@ digraph {
     8 -> 9 [ ]
     8 -> 1 [ ]
     10 -> 11 [ ]
-    11 -> 35 [ ]
-    12 -> 33 [ ]
+    11 -> 36 [ ]
     12 -> 34 [ ]
-    13 -> 16 [ ]
-    13 -> 30 [ ]
-    13 -> 32 [ ]
+    12 -> 35 [ ]
+    13 -> 14 [ ]
+    13 -> 17 [ ]
+    13 -> 31 [ ]
+    13 -> 33 [ ]
     14 -> 1 [ ]
-    14 -> 15 [ ]
-    16 -> 19 [ ]
-    16 -> 21 [ ]
-    17 -> 18 [ ]
-    17 -> 4 [ ]
-    19 -> 17 [ ]
-    20 -> 14 [ ]
-    21 -> 20 [ style = dotted]
-    21 -> 22 [ ]
-    23 -> 0 [ ]
-    23 -> 28 [ ]
-    24 -> 7 [ ]
-    24 -> 15 [ ]
-    25 -> 24 [ ]
-    25 -> 20 [ ]
-    25 -> 26 [ ]
-    25 -> 22 [ ]
-    28 -> 25 [ ]
-    28 -> 27 [ ]
-    28 -> 29 [ ]
-    29 -> 22 [ ]
+    15 -> 1 [ ]
+    15 -> 16 [ ]
+    17 -> 20 [ ]
+    17 -> 22 [ ]
+    18 -> 19 [ ]
+    18 -> 4 [ ]
+    20 -> 18 [ ]
+    21 -> 15 [ ]
+    22 -> 21 [ style = dotted]
+    22 -> 23 [ ]
+    24 -> 0 [ ]
+    24 -> 29 [ ]
+    25 -> 7 [ ]
+    25 -> 16 [ ]
+    26 -> 25 [ ]
+    26 -> 21 [ ]
+    26 -> 27 [ ]
+    26 -> 23 [ ]
+    29 -> 26 [ ]
+    29 -> 28 [ ]
+    29 -> 30 [ ]
     30 -> 23 [ ]
-    30 -> 30 [ color = blue]
-    30 -> 21 [ ]
-    30 -> 31 [ color = blue]
-    31 -> 26 [ ]
-    32 -> 3 [ ]
-    32 -> 22 [ ]
-    33 -> 30 [ ]
-    34 -> 13 [ ]
-    35 -> 12 [ ]
-    35 -> 36 [ ]
-    36 -> 22 [ ]
-    38 -> 10 [ ]
-    39 -> 1 [ ]
-    40 -> 22 [ ]
-    41 -> 42 [ ]
-    42 -> 10 [ ]
+    31 -> 24 [ ]
+    31 -> 31 [ color = blue]
+    31 -> 22 [ ]
+    31 -> 32 [ color = blue]
+    32 -> 27 [ ]
+    33 -> 3 [ ]
+    33 -> 23 [ ]
+    34 -> 31 [ ]
+    35 -> 13 [ ]
+    36 -> 12 [ ]
+    36 -> 37 [ ]
+    37 -> 23 [ ]
+    39 -> 10 [ ]
+    40 -> 1 [ ]
+    41 -> 23 [ ]
+    42 -> 43 [ ]
+    43 -> 10 [ ]
 }
