+++ nomos-services/cryptarchia-consensus/Cargo.toml
@@ -10,8 +10,10 @@ version = "0.1.0"
 async-trait              = "0.1"
 cl                       = { workspace = true }
 cryptarchia-engine       = { workspace = true, features = ["serde"] }
+cryptarchia-sync         = { workspace = true }
 cryptarchia-sync-network = { workspace = true }
 futures                  = "0.3"
+itertools                = "0.13"
 nomos-blend-service      = { workspace = true }
 nomos-core               = { workspace = true }
 nomos-da-sampling        = { workspace = true }
@@ -34,7 +36,6 @@ tokio-stream             = "0.1"
 tracing                  = "0.1"
 tracing-futures          = "0.2"
 utoipa                   = { version = "4.0", optional = true }
-itertools = "0.13.0"
 
 [features]
 default = []
