+++ nomos-services/cryptarchia-consensus/src/network/mod.rs
@@ -2,10 +2,9 @@ pub mod adapters;
 
 use std::hash::Hash;
 
-use cryptarchia_engine::Slot;
 use cryptarchia_sync_network::behaviour::BehaviourSyncReply;
 use futures::Stream;
-use nomos_core::{block::Block, header::HeaderId};
+use nomos_core::block::Block;
 use nomos_network::{
     backends::{libp2p::SyncRequestKind, NetworkBackend},
     NetworkService,
@@ -41,14 +40,6 @@ pub trait NetworkAdapter<RuntimeServiceId> {
 
     async fn blocks_stream(
         &self,
-    ) -> Result<BoxedStream<Block<Self::Tx, Self::BlobCertificate>>, DynError>;
-    async fn fetch_blocks_from_slot(
-        &self,
-        start_slot: Slot,
-    ) -> Result<BoxedStream<Block<Self::Tx, Self::BlobCertificate>>, DynError>;
-    async fn fetch_chain_backward(
-        &self,
-        tip: HeaderId,
     ) -> Result<BoxedStream<Block<Self::Tx, Self::BlobCertificate>>, DynError>;
 
     async fn sync_requests_stream(&self) -> Result<BoxedStream<SyncRequest>, DynError>;
