+++ nomos-services/cryptarchia-consensus/src/sync.rs
@@ -1,190 +1,423 @@
-use std::{
-    collections::{HashMap, HashSet, VecDeque},
-    fmt::Debug,
-    hash::Hash,
-    marker::{PhantomData, Unpin},
-};
+use core::fmt::Debug;
+use std::{hash::Hash, marker::PhantomData};
 
 use cryptarchia_engine::Slot;
-use futures::{Stream, StreamExt};
-use itertools::Itertools;
-use nomos_core::{block::Block, header::HeaderId};
-use serde::{de::DeserializeOwned, Serialize};
+use cryptarchia_sync::adapter::CryptarchiaAdapterError;
+use nomos_core::{
+    block::{AbstractBlock, Block},
+    da::blob::{info::DispersedBlobInfo, metadata::Metadata as BlobMetadata, BlobSelect},
+    header::HeaderId,
+    tx::{Transaction, TxSelect},
+};
+use nomos_da_sampling::backend::DaSamplingServiceBackend;
+use nomos_mempool::{backend::RecoverableMempool, network::NetworkAdapter as MempoolAdapter};
+use nomos_storage::backends::StorageBackend;
+use rand::{RngCore, SeedableRng};
+use serde::{de::DeserializeOwned, Deserialize, Serialize};
+use tokio::sync::broadcast;
 
-use crate::network;
+use crate::{
+    blend, leadership, network::NetworkAdapter, relays::CryptarchiaConsensusRelays, Cryptarchia,
+    CryptarchiaConsensus, Error,
+};
 
-// TODO: Consider moving this to a separate crate.
-// (e.g. consensus/cryptarchia-sync/core)
-
-// TODO: Replace Tx and BlobCertificate with Block if possible.
-// This requires refactoring NetworkAdapter.
-pub struct Synchronization<Cryptarchia, Network, Tx, BlobCertificate, RuntimeServiceId> {
-    _marker: PhantomData<(Cryptarchia, Network, Tx, BlobCertificate, RuntimeServiceId)>,
+#[expect(
+    clippy::type_complexity,
+    reason = "CryptarchiaConsensus amount of generics."
+)]
+pub struct CryptarchiaSyncAdapter<
+    NetAdapter,
+    BlendAdapter,
+    ClPool,
+    ClPoolAdapter,
+    DaPool,
+    DaPoolAdapter,
+    TxS,
+    BS,
+    Storage,
+    SamplingBackend,
+    SamplingNetworkAdapter,
+    SamplingRng,
+    SamplingStorage,
+    DaVerifierBackend,
+    DaVerifierNetwork,
+    DaVerifierStorage,
+    TimeBackend,
+    ApiAdapter,
+    RuntimeServiceId,
+> where
+    NetAdapter: NetworkAdapter<RuntimeServiceId>,
+    NetAdapter::Backend: 'static,
+    NetAdapter::Settings: Send,
+    BlendAdapter: blend::BlendAdapter<RuntimeServiceId>,
+    BlendAdapter::Settings: Send,
+    ClPool: RecoverableMempool<BlockId = HeaderId>,
+    ClPool::RecoveryState: Serialize + for<'de> Deserialize<'de>,
+    ClPool::Settings: Clone,
+    ClPool::Item: Clone + Eq + Hash + Debug + 'static,
+    ClPool::Key: Debug + 'static,
+    ClPoolAdapter: MempoolAdapter<RuntimeServiceId, Payload = ClPool::Item, Key = ClPool::Key>,
+    DaPool: RecoverableMempool<BlockId = HeaderId>,
+    DaPool::RecoveryState: Serialize + for<'de> Deserialize<'de>,
+    DaPool::Item: Clone + Eq + Hash + Debug + 'static,
+    DaPool::Key: Debug + 'static,
+    DaPool::Settings: Clone,
+    DaPoolAdapter: MempoolAdapter<RuntimeServiceId, Key = DaPool::Key>,
+    DaPoolAdapter::Payload: DispersedBlobInfo + Into<DaPool::Item> + Debug,
+    TxS: TxSelect<Tx = ClPool::Item>,
+    TxS::Settings: Send,
+    BS: BlobSelect<BlobId = DaPool::Item>,
+    BS::Settings: Send,
+    Storage: StorageBackend + Send + Sync + 'static,
+    SamplingBackend: DaSamplingServiceBackend<SamplingRng, BlobId = DaPool::Key> + Send + 'static,
+    SamplingBackend::Settings: Clone,
+    SamplingBackend::Share: Debug + 'static,
+    SamplingBackend::BlobId: Debug + 'static,
+    SamplingRng: SeedableRng + RngCore,
+    DaVerifierBackend: nomos_da_verifier::backend::VerifierBackend + Send + 'static,
+    DaVerifierBackend::Settings: Clone,
+{
+    cryptarchia: Option<Cryptarchia>,
+    leader: leadership::Leader,
+    relays: CryptarchiaConsensusRelays<
+        BlendAdapter,
+        BS,
+        ClPool,
+        ClPoolAdapter,
+        DaPool,
+        DaPoolAdapter,
+        NetAdapter,
+        SamplingBackend,
+        SamplingRng,
+        Storage,
+        TxS,
+        DaVerifierBackend,
+        RuntimeServiceId,
+    >,
+    block_broadcaster: broadcast::Sender<Block<ClPool::Item, DaPool::Item>>,
+    _marker: PhantomData<(
+        ApiAdapter,
+        SamplingNetworkAdapter,
+        SamplingStorage,
+        DaVerifierNetwork,
+        DaVerifierStorage,
+        TimeBackend,
+    )>,
 }
 
-impl<Cryptarchia, Network, Tx, BlobCertificate, RuntimeServiceId>
-    Synchronization<Cryptarchia, Network, Tx, BlobCertificate, RuntimeServiceId>
+#[expect(
+    clippy::type_complexity,
+    reason = "CryptarchiaConsensus amount of generics."
+)]
+impl<
+        NetAdapter,
+        BlendAdapter,
+        ClPool,
+        ClPoolAdapter,
+        DaPool,
+        DaPoolAdapter,
+        TxS,
+        BS,
+        Storage,
+        SamplingBackend,
+        SamplingNetworkAdapter,
+        SamplingRng,
+        SamplingStorage,
+        DaVerifierBackend,
+        DaVerifierNetwork,
+        DaVerifierStorage,
+        TimeBackend,
+        ApiAdapter,
+        RuntimeServiceId,
+    >
+    CryptarchiaSyncAdapter<
+        NetAdapter,
+        BlendAdapter,
+        ClPool,
+        ClPoolAdapter,
+        DaPool,
+        DaPoolAdapter,
+        TxS,
+        BS,
+        Storage,
+        SamplingBackend,
+        SamplingNetworkAdapter,
+        SamplingRng,
+        SamplingStorage,
+        DaVerifierBackend,
+        DaVerifierNetwork,
+        DaVerifierStorage,
+        TimeBackend,
+        ApiAdapter,
+        RuntimeServiceId,
+    >
 where
-    Tx: Serialize + DeserializeOwned + Clone + Eq + Hash + Send + 'static,
-    BlobCertificate: Serialize + DeserializeOwned + Clone + Eq + Hash + Send + 'static,
-    Cryptarchia: CryptarchiaAdapter<Tx = Tx, BlobCertificate = BlobCertificate> + Sync + Send,
-    Network: network::NetworkAdapter<RuntimeServiceId, Tx = Tx, BlobCertificate = BlobCertificate>
-        + Sync,
+    NetAdapter: NetworkAdapter<RuntimeServiceId>,
+    NetAdapter::Backend: 'static,
+    NetAdapter::Settings: Send,
+    BlendAdapter: blend::BlendAdapter<RuntimeServiceId>,
+    BlendAdapter::Settings: Send,
+    ClPool: RecoverableMempool<BlockId = HeaderId>,
+    ClPool::RecoveryState: Serialize + for<'de> Deserialize<'de>,
+    ClPool::Settings: Clone,
+    ClPool::Item: Clone + Eq + Hash + Debug + 'static,
+    ClPool::Key: Debug + 'static,
+    ClPoolAdapter: MempoolAdapter<RuntimeServiceId, Payload = ClPool::Item, Key = ClPool::Key>,
+    DaPool: RecoverableMempool<BlockId = HeaderId>,
+    DaPool::RecoveryState: Serialize + for<'de> Deserialize<'de>,
+    DaPool::Item: Clone + Eq + Hash + Debug + 'static,
+    DaPool::Key: Debug + 'static,
+    DaPool::Settings: Clone,
+    DaPoolAdapter: MempoolAdapter<RuntimeServiceId, Key = DaPool::Key>,
+    DaPoolAdapter::Payload: DispersedBlobInfo + Into<DaPool::Item> + Debug,
+    TxS: TxSelect<Tx = ClPool::Item>,
+    TxS::Settings: Send,
+    BS: BlobSelect<BlobId = DaPool::Item>,
+    BS::Settings: Send,
+    Storage: StorageBackend + Send + Sync + 'static,
+    SamplingBackend: DaSamplingServiceBackend<SamplingRng, BlobId = DaPool::Key> + Send + 'static,
+    SamplingBackend::Settings: Clone,
+    SamplingBackend::Share: Debug + 'static,
+    SamplingBackend::BlobId: Debug + 'static,
+    SamplingRng: SeedableRng + RngCore,
+    DaVerifierBackend: nomos_da_verifier::backend::VerifierBackend + Send + 'static,
+    DaVerifierBackend::Settings: Clone,
 {
-    /// Syncs the local block tree with the peers, starting from the local tip.
-    /// This covers the case where the local tip is not on the latest honest
-    /// chain anymore.
-    #[expect(dead_code, reason = "Sync protocol is not integrated yet.")]
-    pub async fn run(
-        mut cryptarchia: Cryptarchia,
-        network: &Network,
-    ) -> Result<Cryptarchia, Box<dyn std::error::Error + Send + Sync + 'static>> {
-        // Repeat the sync process until no peer has a tip ahead of the local tip,
-        // because peers' tips may advance during the sync process.
-        let mut rejected_blocks = HashSet::new();
-        loop {
-            // Fetch blocks from the peers in the range of slots from the local tip to the
-            // latest tip. Gather orphaned blocks, which are blocks from forks
-            // that are absent in the local block tree.
-            let mut orphan_blocks = HashMap::new();
-            let mut num_blocks = 0;
-
-            // TODO: handle network error
-            let mut stream = network
-                .fetch_blocks_from_slot(cryptarchia.tip_slot())
-                .await
-                .unwrap();
-            while let Some(block) = stream.next().await {
-                num_blocks += 1;
-                // Reject blocks that have been rejected in the past
-                // or whose parent has been rejected.
-                let header = block.header().clone();
-                let id = header.id();
-                if rejected_blocks.contains(&id) || rejected_blocks.contains(&header.parent()) {
-                    rejected_blocks.insert(id);
-                    continue;
-                }
-
-                match cryptarchia.process_block(block).await {
-                    Ok(()) => {
-                        orphan_blocks.remove(&id);
-                    }
-                    Err(CryptarchiaAdapterError::ParentNotFound) => {
-                        orphan_blocks.insert(id, header.slot());
-                    }
-                    Err(CryptarchiaAdapterError::InvalidBlock(_)) => {
-                        rejected_blocks.insert(id);
-                    }
-                };
-            }
-
-            // Finish the sync process if no block has been fetched,
-            // which means that no peer has a tip ahead of the local tip.
-            if num_blocks == 0 {
-                break;
-            }
-
-            // Backfill the orphan forks starting from the orphan blocks with applying fork
-            // choice rule. Sort the orphan blocks by slot in descending order
-            // to minimize the number of backfillings.
-            for orphan_block in orphan_blocks
-                .iter()
-                .sorted_by_key(|&(_, slot)| std::cmp::Reverse(slot))
-                .map(|(id, _)| *id)
-            {
-                // Skip the orphan block if it has been processed during the previous
-                // backfillings (i.e. if it has been already added to the local
-                // block tree). Or, skip if it has been rejected during the
-                // previous backfillings.
-                if cryptarchia.has_block(&orphan_block) || rejected_blocks.contains(&orphan_block) {
-                    continue;
-                }
-
-                if let Err((_, invalid_suffix)) =
-                    Self::backfill_fork(&mut cryptarchia, orphan_block, network).await
-                {
-                    rejected_blocks.extend(invalid_suffix);
-                }
-            }
+    pub const fn new(
+        cryptarchia: Cryptarchia,
+        leader: leadership::Leader,
+        relays: CryptarchiaConsensusRelays<
+            BlendAdapter,
+            BS,
+            ClPool,
+            ClPoolAdapter,
+            DaPool,
+            DaPoolAdapter,
+            NetAdapter,
+            SamplingBackend,
+            SamplingRng,
+            Storage,
+            TxS,
+            DaVerifierBackend,
+            RuntimeServiceId,
+        >,
+        block_broadcaster: broadcast::Sender<Block<ClPool::Item, DaPool::Item>>,
+    ) -> Self {
+        Self {
+            cryptarchia: Some(cryptarchia),
+            leader,
+            relays,
+            block_broadcaster,
+            _marker: PhantomData::<(
+                ApiAdapter,
+                SamplingNetworkAdapter,
+                SamplingStorage,
+                DaVerifierNetwork,
+                DaVerifierStorage,
+                TimeBackend,
+            )>,
         }
-
-        Ok(cryptarchia)
     }
 
-    /// Backfills a fork, which is absent in the local block tree
-    /// by fetching blocks from the peers.
-    /// During backfilling, the fork choice rule is continuously applied.
-    async fn backfill_fork(
-        cryptarchia: &mut Cryptarchia,
-        tip: HeaderId,
-        network: &Network,
-    ) -> Result<(), (CryptarchiaAdapterError, Vec<HeaderId>)> {
-        let suffix = Self::find_missing_part(
-            // TODO: handle network error
-            network.fetch_chain_backward(tip).await.unwrap(),
-            cryptarchia,
+    pub fn take(
+        self,
+    ) -> (
+        Cryptarchia,
+        leadership::Leader,
+        CryptarchiaConsensusRelays<
+            BlendAdapter,
+            BS,
+            ClPool,
+            ClPoolAdapter,
+            DaPool,
+            DaPoolAdapter,
+            NetAdapter,
+            SamplingBackend,
+            SamplingRng,
+            Storage,
+            TxS,
+            DaVerifierBackend,
+            RuntimeServiceId,
+        >,
+    ) {
+        (
+            self.cryptarchia.expect("Cryptarchia is not set"),
+            self.leader,
+            self.relays,
         )
-        .await;
-
-        // Add blocks in the fork suffix with applying fork choice rule.
-        // After all, add the tip of the fork suffix to apply the fork choice rule.
-        let mut iter = suffix.into_iter();
-        while let Some(block) = iter.next() {
-            let id = block.header().id();
-            if let Err(e) = cryptarchia.process_block(block).await {
-                return Err((
-                    e,
-                    std::iter::once(id)
-                        .chain(iter.map(|block| block.header().id()))
-                        .collect(),
-                ));
-            };
-        }
-
-        Ok(())
-    }
-
-    /// Finds the point where the fork is disconnected from the local block
-    /// tree, and returns the suffix of the fork from the disconnected point
-    /// to the tip. The disconnected point may be different from the
-    /// divergence point of the fork in the case where the fork has been
-    /// partially backfilled.
-    async fn find_missing_part(
-        mut fork: Box<dyn Stream<Item = Block<Tx, BlobCertificate>> + Send + Sync + Unpin>,
-        cryptarchia: &Cryptarchia,
-    ) -> VecDeque<Block<Tx, BlobCertificate>> {
-        let mut suffix = VecDeque::new();
-        while let Some(block) = fork.next().await {
-            if cryptarchia.has_block(&block.header().id()) {
-                break;
-            }
-            suffix.push_front(block);
-        }
-        suffix
     }
 }
 
 #[async_trait::async_trait]
-pub trait CryptarchiaAdapter {
-    // TODO: Replace Tx and BlobCertificate with Block.
-    //       This requires refactoring NetworkAdapter.
-    type Tx: Clone + Eq + Hash + 'static;
-    type BlobCertificate: Clone + Eq + Hash + 'static;
+impl<
+        NetAdapter,
+        BlendAdapter,
+        ClPool,
+        ClPoolAdapter,
+        DaPool,
+        DaPoolAdapter,
+        TxS,
+        BS,
+        Storage,
+        SamplingBackend,
+        SamplingNetworkAdapter,
+        SamplingRng,
+        SamplingStorage,
+        DaVerifierBackend,
+        DaVerifierNetwork,
+        DaVerifierStorage,
+        TimeBackend,
+        ApiAdapter,
+        RuntimeServiceId,
+    > cryptarchia_sync::adapter::CryptarchiaAdapter
+    for CryptarchiaSyncAdapter<
+        NetAdapter,
+        BlendAdapter,
+        ClPool,
+        ClPoolAdapter,
+        DaPool,
+        DaPoolAdapter,
+        TxS,
+        BS,
+        Storage,
+        SamplingBackend,
+        SamplingNetworkAdapter,
+        SamplingRng,
+        SamplingStorage,
+        DaVerifierBackend,
+        DaVerifierNetwork,
+        DaVerifierStorage,
+        TimeBackend,
+        ApiAdapter,
+        RuntimeServiceId,
+    >
+where
+    NetAdapter: NetworkAdapter<RuntimeServiceId> + Clone + Send + Sync + 'static,
+    NetAdapter::Settings: Send,
+    BlendAdapter: blend::BlendAdapter<RuntimeServiceId> + Clone + Send + Sync + 'static,
+    BlendAdapter::Settings: Send,
+    ClPool: RecoverableMempool<BlockId = HeaderId> + Send + Sync + 'static,
+    ClPool::RecoveryState: Serialize + for<'de> Deserialize<'de>,
+    ClPool::Settings: Clone + Send + Sync + 'static,
+    ClPool::Item: Transaction<Hash = ClPool::Key>
+        + Debug
+        + Clone
+        + Eq
+        + Hash
+        + Serialize
+        + serde::de::DeserializeOwned
+        + Send
+        + Sync
+        + 'static,
+    ClPool::Key: Debug + Send + Sync,
+    ClPoolAdapter: MempoolAdapter<RuntimeServiceId, Payload = ClPool::Item, Key = ClPool::Key>
+        + Send
+        + Sync
+        + 'static,
+    DaPool::Item: DispersedBlobInfo<BlobId = DaPool::Key>
+        + BlobMetadata
+        + Debug
+        + Clone
+        + Eq
+        + Hash
+        + Serialize
+        + DeserializeOwned
+        + Send
+        + Sync
+        + 'static,
+    DaPool: RecoverableMempool<BlockId = HeaderId, Key = SamplingBackend::BlobId>
+        + Send
+        + Sync
+        + 'static,
+    DaPool::RecoveryState: Serialize + for<'de> Deserialize<'de>,
+    DaPool::Settings: Clone + Send + Sync + 'static,
+    DaPoolAdapter: MempoolAdapter<RuntimeServiceId, Key = DaPool::Key> + Send + Sync + 'static,
+    DaPoolAdapter::Payload: DispersedBlobInfo + Into<DaPool::Item> + Debug,
+    TxS: TxSelect<Tx = ClPool::Item> + Clone + Send + Sync + 'static,
+    TxS::Settings: Send,
+    BS: BlobSelect<BlobId = DaPool::Item> + Clone + Send + Sync + 'static,
+    BS::Settings: Send,
+    Storage: StorageBackend + Send + Sync + 'static,
+    SamplingBackend: DaSamplingServiceBackend<SamplingRng> + Send + 'static,
+    SamplingBackend::Settings: Clone,
+    SamplingBackend::Share: Debug + 'static,
+    SamplingBackend::BlobId: Debug + Ord + Send + Sync + 'static,
+    SamplingNetworkAdapter: nomos_da_sampling::network::NetworkAdapter<RuntimeServiceId> + Send,
+    SamplingRng: SeedableRng + RngCore,
+    SamplingStorage: nomos_da_sampling::storage::DaStorageAdapter<RuntimeServiceId> + Send,
+    DaVerifierStorage: nomos_da_verifier::storage::DaStorageAdapter<RuntimeServiceId> + Send + Sync,
+    DaVerifierBackend: nomos_da_verifier::backend::VerifierBackend + Send + Sync + 'static,
+    DaVerifierBackend::Settings: Clone,
+    DaVerifierNetwork: nomos_da_verifier::network::NetworkAdapter<RuntimeServiceId> + Send + Sync,
+    DaVerifierNetwork::Settings: Clone,
+    TimeBackend: nomos_time::backends::TimeBackend + Send,
+    TimeBackend::Settings: Clone + Send + Sync,
+    ApiAdapter: nomos_da_sampling::api::ApiAdapter + Send + Sync,
+{
+    type Block = Block<ClPool::Item, DaPool::Item>;
 
-    async fn process_block(
-        &mut self,
-        block: Block<Self::Tx, Self::BlobCertificate>,
-    ) -> Result<(), CryptarchiaAdapterError>;
+    async fn process_block(&mut self, block: Self::Block) -> Result<(), CryptarchiaAdapterError> {
+        let result = CryptarchiaConsensus::<
+            NetAdapter,
+            BlendAdapter,
+            ClPool,
+            ClPoolAdapter,
+            DaPool,
+            DaPoolAdapter,
+            TxS,
+            BS,
+            Storage,
+            SamplingBackend,
+            SamplingNetworkAdapter,
+            SamplingRng,
+            SamplingStorage,
+            DaVerifierBackend,
+            DaVerifierNetwork,
+            DaVerifierStorage,
+            TimeBackend,
+            ApiAdapter,
+            RuntimeServiceId,
+        >::process_block(
+            self.cryptarchia.take().unwrap(),
+            &mut self.leader,
+            block,
+            &self.relays,
+            &mut self.block_broadcaster,
+        )
+        .await;
 
-    fn tip_slot(&self) -> Slot;
+        match result {
+            Ok(cryptarchia) => {
+                self.cryptarchia = Some(cryptarchia);
+                Ok(())
+            }
+            Err((e, cryptarchia)) => {
+                self.cryptarchia = Some(cryptarchia);
+                match e {
+                    Error::Ledger(nomos_ledger::LedgerError::ParentNotFound(_))
+                    | Error::Consensus(cryptarchia_engine::Error::ParentMissing(_)) => {
+                        Err(CryptarchiaAdapterError::ParentNotFound)
+                    }
+                    e => Err(CryptarchiaAdapterError::InvalidBlock(Box::new(e))),
+                }
+            }
+        }
+    }
 
-    fn has_block(&self, id: &HeaderId) -> bool;
-}
+    fn tip_slot(&self) -> Slot {
+        self.cryptarchia
+            .as_ref()
+            .expect("Cryptarchia shouldn't be in use")
+            .tip_state()
+            .slot()
+    }
 
-#[derive(thiserror::Error, Debug)]
-pub enum CryptarchiaAdapterError {
-    #[error("Parent not found")]
-    ParentNotFound,
-    #[error("Invalid block: {0}")]
-    InvalidBlock(Box<dyn std::error::Error + Send + Sync>),
+    fn has_block(&self, id: &<Self::Block as AbstractBlock>::Id) -> bool {
+        self.cryptarchia
+            .as_ref()
+            .expect("Cryptarchia shouldn't be in use")
+            .ledger
+            .state(id)
+            .is_some()
+    }
 }
